requirements.txt
click
gunicorn
itsdangerous
markupsafe
numpy
pandas
python-dateutil
python-dotenv
pytz 
six
Flask
Jinja2
PyMySQL
SQLAlchemy
Werkzeug